Refactor setUpComponents (#111)
* Factored out setUpActions helpers and sorted actions by task lifecycle

* Moved setUpActions helpers to a separate file
